fix: useless path segments, rename file and Prefer React.Fragment
